Fix a pretty gnarly issue
When the maximum point is not removed from the points list, an infinite loop arises under some circumstances, I haven't figured out the causation. Adding that point to the mesh produces three new faces, one of which is identical to the face which was replaced. The new face can still see the maximum point, and so the process repeats itself. I'm not exactly sure why it is possible to run into this problem, and would need to investigate deeper in order to figure out what face and point characteristics cause it.