Change the way face, vert, and edge ids are generated so they are mesh-specific
They are now mesh-specific, rather than being static local to the program. This makes thinking about and using mesh ids much easier, and removes some unsafe code, but it means that all functions which generate new faces, edges, or vertices for a mesh need to have access to a &mut to the mesh (sometimes this is &mut self). This required re-organizing and refactoring across the codebase, and moving make_triangle out of util.

Note: the Rust compiler came in really handy for doing this refactor, and the borrow system caught an error with a simultaneous read and write that I never would have seen otherwise! It helped me greatly to be able to do essentially a find-and-replace edit to the way objects were created across the class.